Created mock data for Client
