+++ client/src/data/configs.js
@@ -1,0 +1,24 @@
+const configs = [
+    {
+      name: "Config 1",
+      description: "This is the first config",
+      id: 1,
+    },
+    {
+      name: "Config 2",
+      description: "This is the second config",
+      id: 2,
+    },
+    {
+      name: "Config 3",
+      description: "This is the third config",
+      id: 3,
+    },
+    {
+      name: "Config 4",
+      description: "This is the fourth config",
+      id: 4,
+    },
+  ];
+
+export default configs;
