+++ client/src/data/index.js
@@ -1,0 +1,3 @@
+export { default as loadClients } from "./loadClients";
+export { default as configs } from "./configs";
+export { default as loadTests } from "./loadTests";
