+++ client/src/data/loadClients.js
@@ -1,0 +1,37 @@
+const loadClients = [
+    {
+      name: "Load-Client-ID-001",
+      version: 5.0,
+      id: 1,
+      region: "california",
+      zone: "colorado",
+      scheduler: "nev",
+      currstatus: "ready",
+      metrics: "prometheus",
+      dateCreated: 2019,
+    },
+    {
+      name: "Load-Client-ID-002",
+      version: 5.0,
+      id: 2,
+      region: "california",
+      zone: "colorado",
+      scheduler: "nev",
+      currstatus: "unresponsive",
+      metrics: "prometheus",
+      dateCreated: 2021,
+    },
+    {
+      name: "Load-Client-ID-003",
+      version: 5.0,
+      id: 3,
+      region: "california",
+      zone: "colorado",
+      scheduler: "nev",
+      currstatus: "busy",
+      metrics: "prometheus",
+      dateCreated: 2020,
+    },
+  ];
+
+export default loadClients;
