+++ client/src/data/loadTests.js
@@ -1,0 +1,49 @@
+const loadTests = [
+    {
+      name: "Load Test 1",
+      configId: 1,
+      id: 1,
+    },
+    {
+      name: "Load Test 2",
+      configId: 2,
+      id: 2,
+    },
+    {
+      name: "Load Test 3",
+      configId: 1,
+      id: 3,
+    },
+    {
+      name: "Load Test 4",
+      configId: 3,
+      id: 4,
+    },
+    {
+      name: "Load Test 5",
+      configId: 4,
+      id: 5,
+    },
+    {
+      name: "Load Test 6",
+      configId: 2,
+      id: 6,
+    },
+    {
+      name: "Load Test 7",
+      configId: 1,
+      id: 7,
+    },
+    {
+      name: "Load Test 8",
+      configId: 3,
+      id: 8,
+    },
+    {
+      name: "Load Test 9",
+      configId: 1,
+      id: 9,
+    },
+  ];
+
+export default loadTests;
